MAJOR REVISION The goal is to implement nested cross validation for hyperparameter search This involves excluding a run of data for the target subject. Information about this excluded run needed to be encoded in multiple data structures and penetrate multiple functions
We also want to be able to apply specific hyperparameters to model_type/subject/cv combos.
This is handled by adding the `hypfile` command line argument, which expects a path to a json file.
The json file should either be a list of lists: Subjects[CVs[HypCfg]] or a dict of lists where keys are subject labels.

This is all UNTESTED at this point